+++ src/Redball.UI.WPF/ViewModels/MainViewModel.cs
@@ -118,7 +118,6 @@ public class MainViewModel : ViewModelBase
         OpenBehaviorCommand = new RelayCommand(OpenBehavior);
         OpenSmartFeaturesCommand = new RelayCommand(OpenSmartFeatures);
         OpenTypeThingCommand = new RelayCommand(OpenTypeThing);
-        OpenPomodoroCommand = new RelayCommand(OpenPomodoro);
         OpenUpdatesCommand = new RelayCommand(OpenUpdates);
         OpenAboutCommand = new RelayCommand(() => ShowAbout());
         ShowWindowCommand = new RelayCommand(ShowWindow);
@@ -260,7 +259,6 @@ public class MainViewModel : ViewModelBase
     public ICommand OpenBehaviorCommand { get; }
     public ICommand OpenSmartFeaturesCommand { get; }
     public ICommand OpenTypeThingCommand { get; }
-    public ICommand OpenPomodoroCommand { get; }
     public ICommand OpenUpdatesCommand { get; }
     public ICommand OpenAboutCommand { get; }
     public ICommand ShowWindowCommand { get; }
@@ -465,13 +463,6 @@ public class MainViewModel : ViewModelBase
         }
     }
 
-    private void OpenPomodoro()
-    {
-        if (_mainWindowRef != null && _mainWindowRef.TryGetTarget(out var mainWindow))
-        {
-            mainWindow.ShowPomodoro();
-        }
-    }
 
     private void OpenUpdates()
     {